requirements-dev.txt
# Requirements for developers & contributors
# pytest-mock
pytest-asyncio
pytest-httpserver
